+++ airsync-mac/airsync_macApp.swift
@@ -43,15 +43,15 @@ struct airsync_macApp: App {
         }
 
         let rawPortInt = AppState.shared.myDevice?.port ?? Int(Defaults.serverPort)
-        let chosenPort: UInt32
+        let chosenPort: UInt16
         if rawPortInt <= 0 || rawPortInt > 65_535 {
             print("[main-app] Invalid configured port \(rawPortInt). Falling back to 8080.")
-            chosenPort = UInt32(8080)
+            chosenPort = UInt16(8080)
         }
         else {
-            chosenPort = UInt32(rawPortInt)
+            chosenPort = UInt16(rawPortInt)
         }
-        WebSocketServer.shared.start(port: UInt32(chosenPort))
+        WebSocketServer.shared.start(port: UInt16(chosenPort))
 
         Timer.scheduledTimer(withTimeInterval: 10, repeats: true) { _ in
             AppState.shared.syncWithSystemNotifications()
@@ -120,7 +120,7 @@ struct airsync_macApp: App {
                     if let device = appState.device, appState.adbConnected {
                         ADBConnector.startScrcpy(
                             ip: device.ipAddress,
-                            port: UInt32(appState.adbPort),
+                            port: UInt16(appState.adbPort),
                             deviceName: device.name,
                             package: nil
                         )
@@ -137,7 +137,7 @@ struct airsync_macApp: App {
                             if let device = appState.device {
                                 ADBConnector.startScrcpy(
                                     ip: device.ipAddress,
-                                    port: UInt32(appState.adbPort),
+                                    port: UInt16(appState.adbPort),
                                     deviceName: device.name,
                                     package: app.packageName
                                 )
